fix bugs and add response class for the car feature

+++ Company.API/Features/Car/Repositories/CarRepository.cs
@@ -2,62 +2,50 @@
 using CarInformation.API.Features.Car.Repositories.Interface;
 using Microsoft.EntityFrameworkCore;
 using CarInformation.API.Features.Car.Models;
+using CarInformation.API.Features.Car.Dtos;
 
 namespace CarInformation.API.Features.Car.Repositories
 {
     public class CarRepository : ICarRepository
     {
         private readonly CompanyDbContext _context;
+
         public CarRepository(CompanyDbContext context)
         {
             _context = context;
         }
 
-        public async Task AddCarAsync(Features.Car.Models.Car car)
+        public async Task AddCarAsync(Models.Car car)
         {
-            _context.Cars.Add(car);
-            await _context.SaveChangesAsync();
+            await _context.Cars.AddAsync(car);
         }
 
-        public async Task<bool> DeleteCarAsync(int id)
+        public async Task<Models.Car> GetCarByIdAsync(int id)
         {
-            var car = await _context.Cars.FindAsync(id);
-            if (car == null)
-            {
-                return false;
-            }
-
-            _context.Cars.Remove(car);
-            await _context.SaveChangesAsync();
-            return true;
+            return await _context.Cars.FindAsync(id);
         }
 
-        public async Task<IEnumerable<Features.Car.Models.Car>> GetAllCarsAsync()
+        public async Task<IEnumerable<Models.Car>> GetAllCarsAsync()
         {
             return await _context.Cars.ToListAsync();
         }
 
-        public async Task<Features.Car.Models.Car> GetCarByIdAsync(int carId)
+        public async Task UpdateCarAsync(Models.Car car)
         {
-            return await _context.Cars.FindAsync(carId);
+            _context.Cars.Update(car);
         }
 
-        public async Task<IEnumerable<Features.Car.Models.Car>> GetCarsByCompanyAsync(int companyId)
+        public async Task DeleteCarAsync(int id)
         {
-            return await _context.Cars.Where(a => a.CompanyId == companyId).ToListAsync();
+            var car = await GetCarByIdAsync(id);
+            if (car != null) _context.Cars.Remove(car);
         }
 
-        public async Task UpdateCarAsync(Features.Car.Models.Car car)
-        {
-            var existingCar = await _context.Cars.FindAsync(car.Id);
-            if (existingCar != null)
-            {
-                existingCar.Model = car.Model;
-                existingCar.Color = car.Color;
-                existingCar.HighestSpeed = car.HighestSpeed;
-                _context.Entry(existingCar).State = EntityState.Modified;
-                await _context.SaveChangesAsync();
-            }
-        }
+        //public async Task<IEnumerable<Models.Car>> GetCarsByCompanyAsync(int companyId)
+        //{
+        //    return await _context.Cars.Where(a => a.CompanyId == companyId).ToListAsync();
+        //}
+
+
     }
 }
